Split Claude background Auto failure tests to satisfy file_length.
